Create model matrices for model fitting

This function takes the forumlas for each parameter mu,sigma,nu,tau,theta,zeta and creates a list of model matrices mm with items mm$x and mm$s for fixed and smooth terms respectively, with each of those lists being lists of each parameter and their respective model matrices

Usage

create_model_matrices(
  mu.formula = ("response ~ 1"),
  sigma.formula = ("1"),
  nu.formula = ("1"),
  tau.formula = ("1"),
  theta.formula = ("1"),
  zeta.formula = ("1"),
  margin.family = NO(),
  copula.family = "N",
  copula.link = NA,
  dataset = NA,
  quiet_gamlss2 = TRUE
)

Arguments

mu.formula

Formula for the mean parameter of the marginal distribution

sigma.formula

Formula for the sigma parameter of the marginal distribution

nu.formula

Formula for the nu parameter of the marginal distribution

tau.formula

Formula for the tau parameter of the marginal distribution

theta.formula

Formula for the theta parameter of the copula distribution

zeta.formula

Formula for the zeta parameter of the copula distribution

margin.family

Marginal distribution specified as a gamlss family object, e.g. GA(), NO(), PO(), NBI(), etc.

copula.family

Copula distribution code, one of "N", "C", "F", "G", "J", or "t".

copula.link

List of link functions for the copula parameters

Value

Returns a list mm with items mm$x and mm$s for fixed and smooth terms respectively, with each of those lists being lists of each parameter and their respective model matrices